requirements.txt
torch
torchvision
matplotlib
opencv-python
Pillow
tqdm